Move `Arithmetic` into `Operators` lint pass
